requirements.txt
alembic==1.18.3
fastapi==0.128.0
uvicorn==0.40.0
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
python-jose==3.5.0
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.22
python-dotenv==1.2.1
fastapi-mail==1.6.1
email-validator==2.3.0
pydantic==2.12.5
cryptography==46.0.4